+++ gems/aws-partitions/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Fix an issue where services without regionalized endpoints do not resolve to a provided FIPS global region.
+
 1.430.0 (2021-03-01)
 ------------------